V4__Add_Favorite_Table_And_Price_Trigger.sql price type changed to `bigint`

+++ src/main/resources/db/migration/V4__Add_Favorite_Table_And_Price_Trigger.sql
@@ -3,9 +3,9 @@ create table if not exists public.favorites
     id    uuid                                                         not null default gen_random_uuid()
         constraint favorites_pk
             primary key,
-    items uuid[]                                                       not null,
+    items uuid[] not null,
     count smallint generated always as (array_length(items, 1)) stored not null,
-    sum   float(2)                                                     not null default 0.0
+    sum   bigint                                                       not null default 0
 );
 
 create trigger on_favorite_items_trigger
@@ -25,11 +25,11 @@ end;
 $BODY$
     language plpgsql;
 
-create or replace function get_items_sum(new_items uuid[]) returns float(2) as
+create or replace function get_items_sum(new_items uuid[]) returns bigint as
 $BODY$
 declare
     table_record record;
-    total        float(2) = 0;
+    total        bigint = 0;
 begin
     for i in 1 .. array_upper(new_items, 1)
         loop
@@ -51,7 +51,7 @@ create trigger on_items_price_trigger
     when (old.price is distinct from new.price)
 execute procedure update_price_for_lists(new.id, old.price, new.price);
 
-create or replace function update_price_for_lists(product_id uuid, old_price float(2), new_price float(2)) returns trigger as
+create or replace function update_price_for_lists(product_id uuid, old_price bigint, new_price bigint) returns trigger as
 $BODY$
 begin
     update public.favorites
@@ -61,7 +61,7 @@ end;
 $BODY$
     language plpgsql;
 
-create or replace function get_items_price(sum float(2), old_price float(2), new_price float(2)) returns float(2) as
+create or replace function get_items_price(sum bigint, old_price bigint, new_price bigint) returns bigint as
 $BODY$
 begin
     return sum + (new_price - old_price);
